fix: make `Decompiler` run in docker
For whatever reason that is completely unknown to me, as I am not a .net
developer, unless I pass `COMPlus_EnableDiagnostics=0` to `Decompiler`
as an environment variable, I get a cryptic error regarding CoreCLR not
being able to initialize. I unironically found a fix for this while I
was attempting to diagnose it, somehow this env variable makes it not
error out. I won't question it any further.